Auth and guest middleware

+++ routes/web.php
@@ -14,22 +14,21 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
-Route::get('/register', [UserController::class, 'register'])->name('user.register');
-Route::post('/', [UserController::class, 'store'])->name('user.store');
-Route::get('/login', [UserController::class, 'login'])->name('user.login');
-Route::post('/', [UserController::class, 'authenticate'])->name('user.authenticate');
+Route::get('/register', [UserController::class, 'register'])->name('user.register')->middleware('guest');
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/authenticate', [UserController::class, 'authenticate'])->name('user.authenticate');
+Route::post('/store', [UserController::class, 'store'])->name('user.store');
 
-Route::post('/logout', [UserController::class, 'logout'])->name('user.logout');
+Route::post('/logout', [UserController::class, 'logout'])->name('user.logout')->middleware('auth');
 
 
 // LISTING CONTROLLER
 Route::get('/', [ListingController::class, 'index'])->name('listing.index');
-Route::get('/listings/create', [ListingController::class, 'create'])->name('listing.create');
-Route::post('/listings', [ListingController::class, 'store'])->name('listing.store');
+Route::get('/listings/create', [ListingController::class, 'create'])->name('listing.create')->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->name('listing.store')->middleware('auth');
 Route::get('/{id}', [ListingController::class, 'show'])->name('listing.show');
-Route::get('/listings/{id}', [ListingController::class, 'edit'])->name('listing.edit');
-Route::patch('/{id}', [ListingController::class, 'update'])->name('listing.update');
-Route::delete('/{id}', [ListingController::class, 'destroy'])->name('listing.destroy');
+Route::get('/listings/{id}', [ListingController::class, 'edit'])->name('listing.edit')->middleware('auth');
+Route::patch('/{id}', [ListingController::class, 'update'])->name('listing.update')->middleware('auth');
+Route::delete('/{id}', [ListingController::class, 'destroy'])->name('listing.destroy')->middleware('auth');
 
 // USER CONTROLLER
-Route::get('/login', [UserController::class, 'login'])->name('user.login');
